requirements.txt
mysql-connector-python==8.0.31
pandas==1.5.1
streamlit==1.14.0
Pillow==9.3.0
cookiecutter==2.1.1
plotly==5.11.0